requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# Core Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.6.0
httpx>=0.25.0
textual>=0.44.1

# Machine Learning & AI
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.2
huggingface-hub>=0.24.0,<1.0
sentence-transformers>=2.2.2
langchain>=0.0.344
faiss-cpu>=1.7.4

# Scientific Computing
numpy>=1.25.0
scipy>=1.11.3
networkx>=3.2.1
pandas>=2.1.4

# Document Processing
docling>=2.12.0
pypdf>=4.0.1

# Utilities
rich>=13.7.0
ipython>=8.20.0
prompt_toolkit>=3.0.43
tqdm>=4.66.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
websockets>=12.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
